+++ src/io/NetpbmIOBase.cpp
@@ -4,7 +4,7 @@
 using namespace sipl;
 
 // Process the header of both ASCII and Binary files
-std::tuple<size_t, size_t, size_t> NetpbmIOBase::process_header(
+std::tuple<int32_t, int32_t, int32_t> NetpbmIOBase::process_header(
     std::ifstream& stream)
 {
     // Get magic number (must be first string in header)
@@ -19,7 +19,7 @@ std::tuple<size_t, size_t, size_t> NetpbmIOBase::process_header(
         stream >> word;
     }
 
-    size_t image_width = size_t(std::stoul(word));
+    int32_t image_width = int32_t(std::stoul(word));
 
     stream >> word;
     while ('#' == word[0]) {
@@ -27,7 +27,7 @@ std::tuple<size_t, size_t, size_t> NetpbmIOBase::process_header(
         stream >> word;
     }
 
-    size_t image_height = size_t(std::stoul(word));
+    int32_t image_height = int32_t(std::stoul(word));
 
     stream >> word;
     while ('#' == word[0]) {
@@ -35,7 +35,7 @@ std::tuple<size_t, size_t, size_t> NetpbmIOBase::process_header(
         stream >> word;
     }
 
-    size_t maxval = std::stoul(word);
+    int32_t maxval = std::stoul(word);
 
     return std::make_tuple(image_height, image_width, maxval);
 }
